fix sound effect volume
