+++ client/src/three/scenes/Hexception.tsx
@@ -3,7 +3,6 @@ import * as THREE from "three";
 import { TileManager } from "@/dojo/modelManager/TileManager";
 import { SetupResult } from "@/dojo/setup";
 import useUIStore from "@/hooks/store/useUIStore";
-import { dir, soundSelector } from "@/hooks/useUISound";
 import { HexPosition, ResourceMiningTypes, SceneName } from "@/types";
 import { Position } from "@/types/Position";
 import { ResourceIcon } from "@/ui/elements/ResourceIcon";
@@ -26,6 +25,7 @@ import { BuildingPreview } from "../components/BuildingPreview";
 import { SMALL_DETAILS_NAME } from "../components/InstancedModel";
 import { createHexagonShape } from "../geometry/HexagonGeometry";
 import { createPausedLabel } from "../helpers/utils";
+import { playBuildingSound } from "../sound/utils";
 import { BuildingSystemUpdate, RealmSystemUpdate } from "../systems/types";
 import { HexagonScene } from "./HexagonScene";
 import {
@@ -299,13 +299,15 @@ export default class HexceptionScene extends HexagonScene {
       // if not building mode
       const { col: outerCol, row: outerRow } = this.tileManager.getHexCoords();
 
+      const { isSoundOn, effectsLevel } = useUIStore.getState();
+
       if (BUILDINGS_CENTER[0] === hexCoords.col && BUILDINGS_CENTER[1] === hexCoords.row) {
         const building = getComponentValue(
           this.dojo.components.Building,
           getEntityIdFromKeys([BigInt(outerCol), BigInt(outerRow), BigInt(hexCoords.col), BigInt(hexCoords.row)]),
         );
 
-        playBuildingSound(BuildingType[building?.category as keyof typeof BuildingType]);
+        playBuildingSound(BuildingType[building?.category as keyof typeof BuildingType], isSoundOn, effectsLevel);
 
         this.state.setSelectedBuildingHex({
           outerCol,
@@ -320,7 +322,7 @@ export default class HexceptionScene extends HexagonScene {
           getEntityIdFromKeys([BigInt(outerCol), BigInt(outerRow), BigInt(hexCoords.col), BigInt(hexCoords.row)]),
         );
 
-        playBuildingSound(BuildingType[building?.category as keyof typeof BuildingType]);
+        playBuildingSound(BuildingType[building?.category as keyof typeof BuildingType], isSoundOn, effectsLevel);
 
         this.state.setSelectedBuildingHex({
           outerCol,
@@ -684,29 +686,3 @@ export default class HexceptionScene extends HexagonScene {
     });
   }
 }
-
-export const playBuildingSound = (buildingType: BuildingType | undefined) => {
-  const buildingSounds: Partial<Record<BuildingType, string>> = {
-    [BuildingType.Castle]: soundSelector.buildCastle,
-    [BuildingType.WorkersHut]: soundSelector.buildWorkHut,
-    [BuildingType.WatchTower]: soundSelector.buildMageTower,
-    [BuildingType.Storehouse]: soundSelector.buildStorehouse,
-    [BuildingType.Bank]: soundSelector.buildLabor,
-    [BuildingType.FragmentMine]: soundSelector.buildMine,
-    [BuildingType.Barracks]: soundSelector.buildBarracks,
-    [BuildingType.ArcheryRange]: soundSelector.buildArcherRange,
-    [BuildingType.Stable]: soundSelector.buildStables,
-    [BuildingType.Farm]: soundSelector.buildFarm,
-    [BuildingType.FishingVillage]: soundSelector.buildFishingVillage,
-    [BuildingType.Market]: soundSelector.buildMarket,
-  };
-
-  const soundFile =
-    buildingType === undefined
-      ? soundSelector.buildCastle
-      : buildingSounds[buildingType as BuildingType] ?? soundSelector.buildMine;
-
-  const soundPath = dir + soundFile;
-
-  new Audio(soundPath).play();
-};
